opt faucet & fix discord gas decimals
